requirements.txt
openai>=1.0.0
flask>=2.0.0
python-dotenv>=1.0.0
werkzeug>=2.0.0